refactor: extract Algorithm enum to module
